Move schema to its own package and add unit testing
This commit moves the code that handles datatype and table schemas
from the main package to the schema package and adds unit testing
for it.

This commit also adds fake GCS downloader and uploader functions
to both main_test.go and schema_test.go.

There are some readability changes in function and/or variables
names but the core logic hasn't changed.

This commit passes both golangci-lint and go test.  Coverage is
around 90% for both main and schema.