Create directory where data will be written to
Because the order in which containers are started in a pod is
not guaranteed, it's possible that the data directory where
jostler has to watch isn't created by the measurement service yet
causing it to exit.

This commit does the same thing that puhser does to address
the race condition: it creates the directory if it doesn't
already exist.

This commit also simplifies rules pertaining to test files in
.golangci.yml.

Changes passed static analyzers and were tested locally.